refactor: moved SentryInitializer to graph-commons
