+++ graph-commons/src/main/scala/ch/datascience/config/sentry/SentryInitializer.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package ch.datascience.triplesgenerator.init
+package ch.datascience.config.sentry
 
 import cats.MonadError
 import io.sentry.Sentry
