+++ graph-commons/src/test/scala/ch/datascience/interpreters/IOSentryInitializer.scala
@@ -1,0 +1,6 @@
+package ch.datascience.interpreters
+
+import cats.effect.IO
+import ch.datascience.config.sentry.SentryInitializer
+
+abstract class IOSentryInitializer extends SentryInitializer[IO]
