+++ triples-generator/build.sbt
@@ -20,5 +20,4 @@ name := "triples-generator"
 
 libraryDependencies += "ch.qos.logback"  % "logback-classic"    % "1.2.3"
 libraryDependencies += "com.lihaoyi"     %% "ammonite-ops"      % "1.6.7"
-libraryDependencies += "io.sentry"       % "sentry-logback"     % "1.7.22"
 libraryDependencies += "org.apache.jena" % "jena-rdfconnection" % "3.12.0"
