+++ webhook-service/build.sbt
@@ -19,4 +19,3 @@
 name := "webhook-service"
 
 libraryDependencies += "ch.qos.logback" % "logback-classic" % "1.2.3"
-libraryDependencies += "io.sentry"      % "sentry-logback"  % "1.7.22"
